Revert unneeded changes.
