+++ packages/ui/react-ui-stack/src/components/Stack.stories.tsx
@@ -10,8 +10,8 @@ import React, { useRef, useState } from 'react';
 import { Mosaic, type MosaicDropEvent, type MosaicMoveEvent, type MosaicOperation, Path } from '@dxos/react-ui-mosaic';
 import { withTheme } from '@dxos/storybook-utils';
 
-import { Stack } from './Stack';
-import { type StackSectionContent, type StackProps, type StackSectionItem } from './props';
+import { type StackSectionContent, type StackSectionItem } from './Section';
+import { Stack, type StackProps } from './Stack';
 import { FullscreenDecorator } from '../testing/decorators';
 import { TestObjectGenerator } from '../testing/generator';
