+++ packages/ui/react-ui-stack/src/components/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './Stack';
-export * from './props';
